# trace-server/API.yaml
###############################################################################
# Copyright (c) 2017 Ericsson
#
# All rights reserved. This program and the accompanying materials are
# made available under the terms of the Eclipse Public License v1.0 which
# accompanies this distribution, and is available at
# http://www.eclipse.org/legal/epl-v10.html
###############################################################################
swagger: '2.0'
info:
  description: Open source Server for viewing and analyzing any type of logs or traces. Its goal is to provide views, graphs, metrics, and more to help extract useful information from traces, in a way that is more user-friendly and informative than huge text dumps.
  version: 0.0.0
  title: Trace Compass Server
  termsOfService: http://tracecompass.org/
  contact:
    email: tracecompass-dev@eclipse.org
  license:
    name: Eclipse Public License (EPL)
    url: https://www.eclipse.org/legal/epl-v10.html
host: localhost:8080
basePath: /tracecompass
tags:
- name: traces
  description: How to add traces to your server, or print out their content
- name: Views
  description: Find out how to display data
schemes:
- http
paths:
  /traces:
    get:
      tags:
      - traces
      summary: Get the list of traces opened on the server
      operationId: getTraces
      produces:
      - application/json
      responses:
        200:
          description: Returns a list of traces
          schema:
            type: array
            items:
              $ref: '#/definitions/TraceModel'
  /traces/{trace}:
    post:
      tags:
      - traces
      summary: Add a trace to the server and open it.
      description: Add a trace from the server's file system to the trace server,
        and try to open it. Return some base information once opened.
      operationId: addTrace
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      parameters:
      - name: trace
        in: path
        description: The name of the trace in the server
        required: true
        type: string
      - name: path
        in: formData
        description: Path to the trace on the server file system
        required: true
        type: string
      - name: typeID
        in: formData
        description: The trace type's Eclipse extension ID.
        required: false
        type: String
      responses:
        200:
          description: The trace has been successfully added to the trace server.
          schema:
            $ref: '#/definitions/TraceModel'
        406:
          description: "Not acceptable: cannot read this trace type"
        409:
          description: "Conflict: there was already a trace with this name"
        500:
          description: Missing the name and path parameters
    delete:
      tags:
      - traces
      summary: Close a trace and remove it from the server
      operationId: deleteTrace
      produces:
      - application/json
      parameters:
      - name: trace
        in: path
        description: The name of the trace in the server
        required: true
        type: string
      responses:
        200:
          description: The traces was successfully deleted
          schema:
            $ref: '#/definitions/TraceModel'
        204:
          description: There was no trace with this name to close / delete
  /traces/{trace}/eventTable:
    get:
      tags:
      - Views
      - traces
      summary: Get a virtual view of events from a trace to populate the table
      operationId: getEvents
      produces:
      - application/json
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - name: name
        in: path
        description: The name of the trace in the server
        required: true
        type: string
      - name: low
        in: query
        description: Rank of the first event to query
        required: true
        type: integer
        format: int64
        minimum: 0
      - name: size
        in: query
        description: Number of events to query
        required: true
        type: integer
        format: int32
        minimum: 0
      responses:
        200:
          description: Returns an EventView model with a 2D array of strings and metadata
          schema:
            type: array
            items:
              $ref: '#/definitions/EventView'
        400:
          description: Bad request, the top index and size must be larger than 0
          schema:
            type: string
        404:
          description: No such trace
          schema:
            type: string
        500:
          description: Error reading the trace
          schema:
            type: string
    post:
      tags:
      - Views
      - traces
      summary: Get a virtual view of events from a trace to populate the table
      operationId: getEvents
      produces:
      - application/json
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - name: name
        in: path
        description: The name of the trace in the server
        required: true
        type: string
      - name: low
        in: query
        description: Rank of the first event to query, among the filtered events
        required: true
        type: integer
        format: int64
        minimum: 0
      - name: size
        in: query
        description: Number of filtered events to return
        required: true
        type: integer
        format: int32
        minimum: 0
      - name: filters
        in: formData
        type: string
        description: Map of column / aspect names to regular expression filters
        required: false
      responses:
        200:
          description: Returns an EventView model with a 2D array of strings and metadata
          schema:
            type: array
            items:
              $ref: '#/definitions/EventView'
        400:
          description: Bad request, the top index and size must be larger than 0
          schema:
            type: string
        404:
          description: No such trace
          schema:
            type: string
        500:
          description: Error reading the trace
          schema:
            type: string
  /traces/{trace}/DiskIO:
    get:
      tags:
      - Views
      summary: Get the XY series for an analysis
      operationId: getXY
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      parameters:
      - name: trace
        in: path
        description: Name of the trace to query
        required: true
        type: string
      - name: start
        in: query
        description: Start Time of the query
        required: true
        type: integer
        format: int64
      - name: end
        in: query
        description: End Time of the query
        required: true
        type: integer
        format: int64
      - name: nb
        in: query
        description: Number of data points to sample
        required: true
        type: integer
        format: int32
      responses:
      responses:
        200:
          description: Returns a list of view entries
          schema:
            type: array
            items:
              $ref: '#/definitions/XYEntry'
        404:
          description: Trace not found
          schema:
            type: string
        405:
          description: Analysis not possible for this trace
          schema:
            type: string
  /traces/{trace}/DiskIO/xy:
    post:
      tags:
      - Views
      summary: Get the XY series for an analysis
      operationId: getXY
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      parameters:
      - name: trace
        in: path
        description: Name of the trace to query
        required: true
        type: string
      - name: start
        in: query
        description: Start Time of the query
        required: true
        type: integer
        format: int64
      - name: end
        in: query
        description: End Time of the query
        required: true
        type: integer
        format: int64
      - name: nb
        in: query
        description: Number of data points to sample
        required: true
        type: integer
        format: int32
      - name: ids
        in: formData
        description: Parameters to query specific entries
        required: true
        type: array
        items:
          type: integer
          format: int64
      responses:
        200:
          description: Return a list of AnalysisModel
          schema:
            $ref: '#/definitions/XYView'
        404:
          description: Trace not found
          schema:
            type: string
        405:
          description: Analysis not possible for this trace
          schema:
            type: string
definitions:
  TraceModel:
    type: object
    properties:
      name:
        description: User defined name for the trace
        type: string
      path:
        description: Path to the trace on the file server's File System
        type: string
      nbEvents:
        description: Current number of indexed events in the trace.
        type: integer
        format: int64
      aspects:
        description: List of the trace's aspect names
        type: array
        items:
          type: string
  EventView:
    type: object
    properties:
      trace:
        $ref: '#/definitions/TraceModel'
      low:
        description: Rank of the first returned event
        type: integer
        format: int64
      size:
        description: Number of events.
        type: integer
        format: int32
      filteredSize:
        description: Total number of events that currently match this filter
        type: integer
        format: int64
      filters:
        description: Map of column / aspect name to regular expression filter
        type: object
        properties:
          key:
            type: string
          value:
            type: string
      lines:
        type: array
        items:
          type: array
          items:
            type: string
  AnalysisModel:
    type: object
    properties:
      name:
        type: string
        description: Name of the analyzed trace.
      start:
        type: integer
        format: int64
        description: Start time of the analysis in nanoseconds
      end:
        type: integer
        format: int64
        description: Current end time of the analysis in nanoseconds
  XYEntry:
    type: object
    properties:
      name:
        description: Displayed name for this Entry
        type: string
      id:
        description: Unique id to identify this entry in the backend.
        type: integer
        format: int32
      parentId:
        description: Unique id to identify this parent's entry.
        type: integer
        format: int32
  XYView:
    type: object
    properties:
      trace:
        description: Metadata on the current Analysis
        $ref: '#/definitions/AnalysisModel'
      model:
        type: object
        properties:
          times:
            description: X values (times) for the XY Model
            type: array
            items:
              type: integer
              format: int64
          values:
            description: Map of Series names to the Y values' arrays.
            type: object
            properties:
              key:
                type: string
              value:
                type: array
                items:
                  type: integer
                  format: int64